[feat] create group & allocation files
- 그룹만들어서 파일할당해줌
