+++ NaverMapA/NaverMapA/Model/Cluster/Cluster.swift
@@ -1,0 +1,14 @@
+//
+//  Cluster.swift
+//  NaverMapA
+//
+//  Created by 채훈기 on 2020/11/26.
+//
+
+import Foundation
+
+protocol Cluster {
+    var latitude: Double { get set }
+    var longitude: Double { get set }
+    var places: [Place] { get set }
+}
