+++ NaverMapA/NaverMapA/Model/Place/JsonPlace.swift
