+++ NaverMapA/NaverMapA/Model/Place/NaverMapA.xcdatamodeld/NaverMapA.xcdatamodel/contents
